pyproject.toml
[tool.poetry]
name = "ivcap_ai_tool"
version = "0.7.18"
description = "Helper functions for building AI Tools as IVCAP services"

authors = ["Max Ott <max.ott@csiro.au>"]

readme = "README.md"

include = ["ivcap_ai_tool/py.typed"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
uuid6 = "2024.7.10"
cachetools = "^5.5.2"
opentelemetry-instrumentation-fastapi = "^0.51b0"
ivcap-service = "^0.6.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.1.0"
Sphinx = "^5.2.3"
# myst-nb = "^1.2.0"
autoapi = "^2.0.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.0.0"
licenseheaders = "^0.8.8"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
